+++ R/tsfeatures.R
@@ -4,15 +4,13 @@
 #'
 #' @docType package
 #' @name tsfeatures
-#' @importFrom stats as.ts bw.nrd0 coef dnorm embed fitted frequency lm
+#' @importFrom stats as.ts bw.nrd0 coef dnorm embed fitted frequency lm spec.ar
 #' @importFrom stats median na.contiguous na.pass residuals cor sd tsp "tsp<-" var
 #' @importFrom stats quantile acf pacf stl pchisq ar Box.test poly start cmdscale
 #' @importFrom purrr map map_dbl
 #' @importFrom forecast mstl
 NULL
 # > NULL
-
-
 
 #' Convert mts object to list of time series
 #' @method as.list mts