+++ NEWS.md
@@ -1,3 +1,9 @@
+# tsfeatures (development version)
+
+## Bug fixes
+
+* Better handling of perfect fits in `arch_stat()`
+
 # tsfeatures 1.0.1
 
 * Bug fixes